pyproject.toml
[project]
name = "vibe-voice-vendor"
version = "0.1.0"
description = "Secure queue-based ASR server wrapping VibeVoice-ASR-7B via vLLM"
license = "MIT OR Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    "pydantic>=2.6.0",
    "python-multipart>=0.0.18",
]

[project.scripts]
vvv = "client.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server", "client", "scripts"]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[tool.mypy]
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["personal_server"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
